描画パイプラインをDAG化＆ScreenDrawer/エフェクト統合
描画パイプラインを依存関係付きDAG方式にリファクタリングし、描画順序を自動解決するよう変更。ScreenDrawerクラスを新規追加し、アウトライン・ヴィネット・ガウシアンブラー・グレースケール等のポストエフェクトを統合管理。プレイヤーのダメージ時に画面効果を発動する仕組みを実装。各シェーダーもパラメータ可変化や2パス化など機能拡張。GameScenePhaseの描画・更新もScreenDrawer経由に変更。その他、const化・型修正・不要なEnumToString削除等のリファクタリングを実施。

+++ NobleEngine/src/engine/Window/WindowManager.cpp
@@ -139,7 +139,7 @@ void WindowManager::RegisterMouseRawInput(HWND hwnd)
     ::RegisterRawInputDevices(&rid, 1, sizeof(rid));
 }
 
-void WindowManager::CreateMainWindow(uint32_t width, uint32_t height, const std::wstring& title) 
+void WindowManager::CreateMainWindow(LONG width, LONG height, const std::wstring& title)
 {
     // ウィンドウサイズを表す構造体にクライアント領域を入れる
     RECT wrc = { 0,0,width,height };
